refactor: move standard fields into GenericSectionInterface
