+++ src/schema/v2/homeView/index.ts
@@ -11,7 +11,7 @@ import {
   connectionWithCursorInfo,
   paginationResolver,
 } from "../fields/pagination"
-import { HomeViewGenericSectionInterface } from "./sectionTypes/HomeViewGenericSectionInterface"
+import { HomeViewGenericSectionInterface } from "./sectionTypes/GenericSectionInterface"
 import { getSectionsForUser } from "./helpers/getSectionsForUser"
 import { registry } from "./sections"
 import { isSectionDisplayable } from "./helpers/isSectionDisplayable"
