+++ src/schema/v2/homeView/sectionTypes/HomeViewCardsSectionType.ts
@@ -6,9 +6,9 @@ import {
   emptyConnection,
 } from "../../fields/pagination"
 import { NodeInterface } from "../../object_identification"
-import { HomeViewGenericSectionInterface } from "./HomeViewGenericSectionInterface"
+import { HomeViewGenericSectionInterface } from "./GenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "./standardSectionFields"
+import { standardSectionFields } from "./GenericSectionInterface"
 import Image from "schema/v2/image"
 
 export const HomeViewCardType = new GraphQLObjectType<any, ResolverContext>({
