+++ src/schema/v2/homeView/sectionTypes/HomeViewGalleriesSectionType.ts
@@ -1,9 +1,9 @@
 import { GraphQLObjectType } from "graphql"
 import { ResolverContext } from "types/graphql"
 import { NodeInterface } from "../../object_identification"
-import { HomeViewGenericSectionInterface } from "./HomeViewGenericSectionInterface"
+import { HomeViewGenericSectionInterface } from "./GenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "./standardSectionFields"
+import { standardSectionFields } from "./GenericSectionInterface"
 
 export const HomeViewGalleriesSectionType = new GraphQLObjectType<
   any,
