+++ src/schema/v2/homeView/sectionTypes/HomeViewMarketingCollectionsSectionType.ts
@@ -4,9 +4,9 @@ import { ResolverContext } from "types/graphql"
 import { connectionWithCursorInfo } from "../../fields/pagination"
 import { MarketingCollectionType } from "../../marketingCollections"
 import { NodeInterface } from "../../object_identification"
-import { HomeViewGenericSectionInterface } from "./HomeViewGenericSectionInterface"
+import { HomeViewGenericSectionInterface } from "./GenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "./standardSectionFields"
+import { standardSectionFields } from "./GenericSectionInterface"
 
 export const HomeViewMarketingCollectionsSectionType = new GraphQLObjectType<
   any,
